Ensure that fload does not have alignment issues
Also fixes a bug with 16-bit endian_flip which was never used
but definitely did not work as expected.

+++ lib/include/ert/util/util.hpp
@@ -163,7 +163,7 @@ void *util_malloc(size_t);
 void *util_calloc(size_t elements, size_t element_size);
 void *util_alloc_copy(const void *, size_t);
 char *util_fread_alloc_file_content(const char *, int *);
-void util_endian_flip_vector(void *data, int element_size, int elements);
+void util_endian_flip_vector(void *data, size_t element_size, size_t elements);
 
 void util_double_vector_max_min(int, const double *, double *, double *);
 void util_update_double_max_min(double, double *, double *);
